refactored and restructured code

+++ tcp/common-utils/common.c
@@ -8,10 +8,10 @@
  * whether the string contains the correct number or not, because in
  * cases when argv = 0 and in error cases it will return 0.
  */
-int exclude_port(int argc, char *argv[]) {
+uint16_t exclude_port(int argc, char *argv[]) {
     if (argc != 2) raise_error(WRONG_ARGS_NUMBER);
     char *p;
     int port = (int) strtol(argv[1], &p, DECIMAL);
     if (port <= 0) raise_error(WRONG_PORT_FORMAT);
-    return port;
+    return (uint16_t) port;
 }
